Renamed files to correspond to ModelConfig class name
