+++ ludwig/utils/defaults.py
@@ -23,7 +23,7 @@ from ludwig.contrib import add_contrib_callback_args
 from ludwig.features.feature_registries import input_type_registry
 from ludwig.globals import LUDWIG_VERSION
 from ludwig.schema import validate_config
-from ludwig.schema.config_object import ModelConfig
+from ludwig.schema.model_config import ModelConfig
 from ludwig.schema.preprocessing import PreprocessingConfig
 from ludwig.utils.backward_compatibility import upgrade_config_dict_to_latest_version
 from ludwig.utils.data_utils import load_config_from_str, load_yaml
