+++ tests/integration_tests/test_hyperopt_ray_horovod.py
@@ -27,7 +27,7 @@ from ludwig.globals import HYPEROPT_STATISTICS_FILE_NAME
 from ludwig.hyperopt.results import HyperoptResults
 from ludwig.hyperopt.run import hyperopt
 from ludwig.hyperopt.utils import update_hyperopt_params_with_defaults
-from ludwig.schema.config_object import ModelConfig
+from ludwig.schema.model_config import ModelConfig
 from tests.integration_tests.utils import binary_feature, create_data_set_to_use, generate_data, number_feature
 
 try:
